KAMANJA-435 - KAMANJA 1035
