+++ trunk/KamanjaManager/src/main/scala/com/ligadata/KamanjaManager/KamanjaMetadata.scala
@@ -41,7 +41,6 @@ import com.ligadata.MetadataAPI.MetadataAPIImpl
 import java.util.concurrent.locks.ReentrantReadWriteLock
 
 import com.ligadata.Utils.{KamanjaClassLoader, KamanjaLoaderInfo, Utils}
-import com.ligadata.python.PyServerConnection
 import org.json4s._
 import org.json4s.native.JsonMethods._
 
@@ -1163,31 +1162,7 @@ object KamanjaMetadata extends ObjectResolver {
               case "Add" => {
                 try {
                   val mdl = mdMgr.Model(zkMessage.NameSpace, zkMessage.Name, zkMessage.Version.toLong, true)
-                  if (mdl != None) {
-		     val model : ModelDef = mdl.get
-		     if (model.miningModelType == MiningModelType.PYTHON) {
-		        LOG.debug("Something is happening to the model here blah blah blah ------" + model.objectDefinition)
-  	             	val connectionMap : scala.collection.mutable.HashMap[String,Any] =
-                        txnCtxt.nodeCtxt.getValue("PYTHON_CONNECTIONS").asInstanceOf[scala.collection.mutable.HashMap[String,Any]]
-			LOG.debug("Something is happening to the model here Haha Haha Haha ------" + connectionMap.size)
-	             	val modelOptStr : String = model.modelConfig
-                        val trimmedOptionStr : String = modelOptStr.trim
-                        val modelOptions : Map[String,Any] = parse(trimmedOptionStr).values.asInstanceOf[Map[String,Any]]
-		        val moduleModelNms : Array[String] = model.PhysicalName.split('.')
-		        val moduleName : String = moduleModelNms.head
-		        val modelName : String = moduleModelNms.last
-			val modelPath : String = txnCtxt.nodeCtxt.getValue("PYTHON_PATH") + "/models/" + moduleName + ".py"
-			writeSrcFile( model.objectDefinition, modelPath) 
-			for ((partitionKey, connection) <- connectionMap) {
-			  val pySrvConn : PyServerConnection = connectionMap.getOrElse(partitionKey, null).asInstanceOf[PyServerConnection]
-			   if (pySrvConn != null) {
-			      val resultStr : String = pySrvConn.addModel(moduleName, modelName, model.objectDefinition, modelOptions)
-			   }
-			}
-			
-			}
-  		     allJarsToBeValidated ++= GetAllJarsFromElem(mdl.get)
-                  }
+  		  allJarsToBeValidated ++= GetAllJarsFromElem(mdl.get)
                 } catch {
                   case e: Exception => {
                     LOG.debug("", e)
